Move filename-gathering method to IOUtilities
