+++ AudioTagger.Library/AudioTagger.Library.csproj
@@ -5,11 +5,14 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="TagLibSharp" Version="2.3.0"/>
-    <PackageReference Include="System.Configuration.ConfigurationManager" Version="7.0.0"/>
-    <PackageReference Include="System.Drawing.Common" Version="7.0.0"/>
-    <PackageReference Include="Microsoft.Windows.Compatibility" Version="7.0.3"/>
-    <PackageReference Include="runtime.osx.10.10-x64.CoreCompat.System.Drawing" Version="6.0.5.128"/>
-    <PackageReference Include="FluentResults" Version="3.15.2"/>
+    <PackageReference Include="TagLibSharp" Version="2.3.0" />
+    <PackageReference Include="System.Configuration.ConfigurationManager" Version="7.0.0" />
+    <PackageReference Include="System.Drawing.Common" Version="7.0.0" />
+    <PackageReference Include="Microsoft.Windows.Compatibility" Version="7.0.3" />
+    <PackageReference Include="runtime.osx.10.10-x64.CoreCompat.System.Drawing" Version="6.0.5.128" />
+    <PackageReference Include="FluentResults" Version="3.15.2" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\AudioTagger.Library.FSharp\AudioTagger.Library.FSharp.fsproj" />
   </ItemGroup>
 </Project>
